requirements.txt
# Deep Learning Framework
torch
torchvision

# Image Processing
opencv-python
pillow
scikit-image
lpips  # Learned Perceptual Image Patch Similarity

# Scientific Computing
numpy
scipy

# Machine Learning Utilities
wandb  # Weights & Biases for experiment tracking

# Deployment
fastapi
uvicorn

# TensorRT Integration
tensorrt

# Utilities
tqdm
pyyaml
matplotlib
exifread
hydra
omegaconf
hydra-core

# Optional: CUDA Support (uncomment if needed)
# cupy-cuda11x  # Replace 11x with your CUDA version
onnx
onnxruntime-gpu
pycuda
